active-learning-esp-feat-sequential-query/active-learning-esp-feat-sequential-query/requirements.txt
adabelief-pytorch>=0.1.0
matplotlib>=3.3.1
modAL>=0.4.0
numpy>=1.19.4
pandas>=1.1.0
scikit-learn>=0.23.2
scipy>=1.4.1
seaborn>=0.11.0
skorch>=0.9.0
torch>=1.6.0
torchvision>=0.7.0
tqdm>=4
PyYAML>=5
tripletnet-classifier @ git+https://github.com/Lucashsmello/TripletNet-on-ESP
#pip install git+http://gitlab.ninfa.inf.ufes.br/ninfa-ufes/deep-rpdbcs#subdirectory=src/python
